# packages/openapi/agent-api.yaml
openapi: 3.1.0
info:
  title: HashHive Agent API
  description: Token-authenticated REST API for Go-based hashcat agents
  version: 1.0.0

servers:
  - url: /api/v1/agent
    description: Agent API base path

paths:
  /sessions:
    post:
      operationId: createAgentSession
      summary: Authenticate agent and obtain session token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Agent auth token
                  example: "agent-secret-token-abc123"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionToken:
                    type: string
                    description: JWT session token for subsequent requests
                  config:
                    type: object
                    properties:
                      agentId:
                        type: integer
                      projectId:
                        type: integer
        "401":
          $ref: "#/components/responses/AuthError"

  /heartbeat:
    post:
      operationId: sendHeartbeat
      summary: Send agent heartbeat with status and capabilities
      tags: [Monitoring]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Heartbeat"
      responses:
        "200":
          $ref: "#/components/responses/Acknowledged"
        "401":
          $ref: "#/components/responses/AuthError"

  /tasks/next:
    post:
      operationId: getNextTask
      summary: Request the next available task matching agent capabilities
      tags: [Tasks]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Next task or null if none available
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    oneOf:
                      - $ref: "#/components/schemas/TaskDescriptor"
                      - type: "null"
        "401":
          $ref: "#/components/responses/AuthError"

  /tasks/{taskId}/report:
    post:
      operationId: reportTaskProgress
      summary: Report task progress, results, and errors
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskReport"
      responses:
        "200":
          $ref: "#/components/responses/Acknowledged"
        "401":
          $ref: "#/components/responses/AuthError"

  /tasks/{taskId}/zaps:
    get:
      operationId: getTaskZaps
      summary: Retrieve cracked hash values for a task
      description: >
        Returns hash values that have been cracked from the same hash list as the
        given task. Agents use this to build a "zap list" so they can skip
        already-cracked hashes during processing.
      tags: [Tasks]
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: since
          in: query
          required: false
          description: Only return hashes cracked after this ISO-8601 datetime
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          description: Maximum number of zaps to return (1-10000, default 10000)
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 10000
      responses:
        "200":
          description: List of cracked hash values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZapResponse"
        "401":
          $ref: "#/components/responses/AuthError"
        "404":
          description: Task not found or not assigned to this agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /benchmark:
    post:
      operationId: submitBenchmarks
      summary: Submit hashcat benchmark results for this agent
      tags: [Monitoring]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BenchmarkSubmission"
      responses:
        "200":
          $ref: "#/components/responses/Acknowledged"
        "401":
          $ref: "#/components/responses/AuthError"

  /errors:
    post:
      operationId: reportAgentError
      summary: Report an agent error
      tags: [Monitoring]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentError"
      responses:
        "200":
          $ref: "#/components/responses/Acknowledged"
        "401":
          $ref: "#/components/responses/AuthError"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token

  schemas:
    Heartbeat:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [online, busy, error]
        capabilities:
          type: object
          properties:
            hashcatVersion:
              type: string
            gpuDevices:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  memory:
                    type: number
                  computeCapability:
                    type: string
        deviceInfo:
          type: object
          properties:
            cpuUsage:
              type: number
            memoryUsage:
              type: number
            temperature:
              type: number

    TaskDescriptor:
      type: object
      properties:
        id:
          type: integer
        attackId:
          type: integer
        campaignId:
          type: integer
        mode:
          type: integer
          description: hashcat attack mode
        hashTypeId:
          type: integer
        workRange:
          type: object
          properties:
            start:
              type: integer
            end:
              type: integer
            agentSpeedHs:
              type: integer
              minimum: 0
              description: Agent benchmark speed in hashes/second for the task's hash mode (fallback 1 MH/s if no benchmark)
        resources:
          type: object
          properties:
            hashListUrl:
              type: string
              format: uri
            wordlistUrl:
              type: string
              format: uri
            rulelistUrl:
              type: string
              format: uri
            masklistUrl:
              type: string
              format: uri

    TaskReport:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [running, completed, failed, exhausted]
        progress:
          type: object
          properties:
            keyspaceProgress:
              type: number
            speed:
              type: number
            temperature:
              type: number
        results:
          type: array
          description: Batch of cracked hash results
          items:
            type: object
            required: [hashValue, plaintext]
            properties:
              hashValue:
                type: string
              plaintext:
                type: string
        errors:
          type: array
          items:
            type: string

    AgentError:
      type: object
      required: [severity, message]
      properties:
        severity:
          type: string
          enum: [warning, error, fatal]
        message:
          type: string
        context:
          type: object
          additionalProperties: true
        taskId:
          type: integer

    ZapResponse:
      type: object
      required: [zaps, hasMore]
      properties:
        zaps:
          type: array
          description: Cracked hash values from the task's hash list
          items:
            type: string
        hasMore:
          type: boolean
          description: Whether more zaps are available beyond the requested limit

    BenchmarkSubmission:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          minItems: 1
          items:
            type: object
            required: [hashcatMode, hashType, speedHs, deviceName]
            properties:
              hashcatMode:
                type: integer
                minimum: 0
              hashType:
                type: string
              speedHs:
                type: integer
                minimum: 0
                description: Hashes per second
              deviceName:
                type: string
        crackerVersion:
          type: string
          minLength: 1
          description: Hashcat version string (e.g. "6.2.6")

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

  responses:
    Acknowledged:
      description: Request acknowledged
      content:
        application/json:
          schema:
            type: object
            properties:
              acknowledged:
                type: boolean
                example: true

    AuthError:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"